requirements.txt
asgiref==3.6.0
certifi==2022.12.7
charset-normalizer==3.0.1
Django==4.1.7
djangorestframework==3.14.0
greenlet==2.0.2
idna==3.4
numpy==1.24.2
openpyxl==3.1.2
pandas==1.5.3
pyodbc==4.0.35
python-dateutil==2.8.2
pytz==2022.7.1
requests==2.28.2
six==1.16.0
spring-config-client==0.2
SQLAlchemy==2.0.5.post1
sqlparse==0.4.3
typing_extensions==4.5.0
tzdata==2022.7
urllib3==1.26.14
XlsxWriter==3.0.8